Fix export file dates and cleanup date util file names
https://issues.redhat.com/browse/COST-3149
