+++ src/routes/views/components/groupBy/groupBy.tsx
@@ -6,7 +6,8 @@ import messages from 'locales/messages';
 import React from 'react';
 import { injectIntl, WrappedComponentProps } from 'react-intl';
 import { connect } from 'react-redux';
-import { getDateRange, getDateRangeDefault, PerspectiveType } from 'routes/views/explorer/explorerUtils';
+import { PerspectiveType } from 'routes/views/explorer/explorerUtils';
+import { getDateRange, getDateRangeDefault } from 'routes/views/utils/dateRange';
 import { createMapStateToProps, FetchStatus } from 'store/common';
 import { orgActions, orgSelectors } from 'store/orgs';
 import { tagActions, tagSelectors } from 'store/tags';
